initial changes to home/news screen and integration of backend for both os
