+++ android/app/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.pintsize.pintsize">
+    package="com.example.pintsize">
    <application
         android:label="pintsize"
         android:icon="@mipmap/ic_launcher">
