+++ ios/.gitignore
@@ -25,7 +25,8 @@ Flutter/flutter_assets/
 Flutter/flutter_export_environment.sh
 ServiceDefinitions.json
 Runner/GeneratedPluginRegistrant.*
-
+Runner/GoogleService-Info.plist
+GoogleService-Info.plist
 # Exceptions to above rules.
 !default.mode1v3
 !default.mode2v3
