+++ ios/Podfile
@@ -1,10 +1,10 @@
-# Uncomment the next line to define a global platform for your project
-platform :ios, '10.0'
+# Uncomment this line to define a global platform for your project
+# platform :ios, '9.0'
 
+# CocoaPods analytics sends network stats synchronously affecting flutter build latency.
 ENV['COCOAPODS_DISABLE_STATS'] = 'true'
 
 project 'Runner', {
-
   'Debug' => :debug,
   'Profile' => :release,
   'Release' => :release,
@@ -31,12 +31,8 @@ target 'Runner' do
   use_frameworks!
   use_modular_headers!
 
-
-
   flutter_install_all_ios_pods File.dirname(File.realpath(__FILE__))
 end
-
-
 
 post_install do |installer|
   installer.pods_project.targets.each do |target|
