--- ios/Runner/GoogleService-Info.plist
@@ -1,38 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>CLIENT_ID</key>
-	<string>535663490773-814a1fdo3aimbo8djvttojncikaehq6l.apps.googleusercontent.com</string>
-	<key>REVERSED_CLIENT_ID</key>
-	<string>com.googleusercontent.apps.535663490773-814a1fdo3aimbo8djvttojncikaehq6l</string>
-	<key>ANDROID_CLIENT_ID</key>
-	<string>535663490773-joukjo84a9ghgnot6b62ohut1habt32p.apps.googleusercontent.com</string>
-	<key>API_KEY</key>
-	<string>AIzaSyAdtMH9oobDAe5F2BTNF3wc7a0n5WjqVH0</string>
-	<key>GCM_SENDER_ID</key>
-	<string>535663490773</string>
-	<key>PLIST_VERSION</key>
-	<string>1</string>
-	<key>BUNDLE_ID</key>
-	<string>com.example.pintsize.pintsize</string>
-	<key>PROJECT_ID</key>
-	<string>pintsize-5c61a</string>
-	<key>STORAGE_BUCKET</key>
-	<string>pintsize-5c61a.appspot.com</string>
-	<key>IS_ADS_ENABLED</key>
-	<false></false>
-	<key>IS_ANALYTICS_ENABLED</key>
-	<false></false>
-	<key>IS_APPINVITE_ENABLED</key>
-	<true></true>
-	<key>IS_GCM_ENABLED</key>
-	<true></true>
-	<key>IS_SIGNIN_ENABLED</key>
-	<true></true>
-	<key>GOOGLE_APP_ID</key>
-	<string>1:535663490773:ios:1f400fbc7171cbddd2145c</string>
-	<key>DATABASE_URL</key>
-	<string>https://pintsize-5c61a.firebaseio.com</string>
-</dict>
-</plist>
