+++ lib/Config/backgroundPalette.dart
