--- lib/Screens/Home/homeTabs/mythsFacts.dart
@@ -1,89 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:pintsize/Widgets/bottomNav/bottom_bar.dart';
-
-class HomeMythsnFactsSubPage extends StatelessWidget {
-  final assetPath, cookieprice, cookiename;
-
-  HomeMythsnFactsSubPage({this.assetPath, this.cookieprice, this.cookiename});
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        backgroundColor: Colors.white,
-        elevation: 0.0,
-        centerTitle: true,
-        leading: IconButton(
-          icon: Icon(Icons.arrow_back, color: Color(0xFF545D68)),
-          onPressed: () {
-            Navigator.of(context).pop();
-          },
-        ),
-        title: Text('Card Title',
-            style: TextStyle(
-                fontFamily: 'Varela',
-                fontSize: 20.0,
-                color: Color(0xFF545D68))),
-        actions: <Widget>[
-          IconButton(
-            icon: Icon(Icons.notifications_none, color: Color(0xFF545D68)),
-            onPressed: () {},
-          ),
-        ],
-      ),
-      body: ListView(children: [
-        SizedBox(height: 15.0),
-        Padding(
-          padding: EdgeInsets.only(left: 20.0),
-          child: Text('Title',
-              style: TextStyle(
-                  fontFamily: 'Varela',
-                  fontSize: 42.0,
-                  fontWeight: FontWeight.bold,
-                  color: Color(0xFFF17532))),
-        ),
-        SizedBox(height: 15.0),
-        Hero(
-            tag: assetPath,
-            child: Image.asset(assetPath,
-                height: 150.0, width: 100.0, fit: BoxFit.contain)),
-        SizedBox(height: 20.0),
-        Center(
-          child: Text(cookieprice,
-              style: TextStyle(
-                  fontFamily: 'Varela',
-                  fontSize: 22.0,
-                  fontWeight: FontWeight.bold,
-                  color: Color(0xFFF17532))),
-        ),
-        SizedBox(height: 10.0),
-        Center(
-          child: Text(cookiename,
-              style: TextStyle(
-                  color: Color(0xFF575E67),
-                  fontFamily: 'Varela',
-                  fontSize: 24.0)),
-        ),
-        SizedBox(height: 20.0),
-        Center(
-          child: Container(
-            width: MediaQuery.of(context).size.width - 50.0,
-            child: Text('Details/Data',
-                textAlign: TextAlign.center,
-                style: TextStyle(
-                    fontFamily: 'Varela',
-                    fontSize: 16.0,
-                    color: Color(0xFFB4B8B9))),
-          ),
-        ),
-        SizedBox(height: 20.0),
-      ]),
-      floatingActionButton: FloatingActionButton(
-        onPressed: () {},
-        backgroundColor: Color(0xFFF17532),
-        child: Icon(Icons.fast_forward),
-      ),
-      floatingActionButtonLocation: FloatingActionButtonLocation.centerDocked,
-      bottomNavigationBar: BottomBar(),
-    );
-  }
-}
