+++ lib/main.dart
@@ -1,13 +1,13 @@
-import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
+import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/services.dart';
-import 'package:shared_preferences/shared_preferences.dart';
+import 'package:pintsize/Screens/News/news.dart';
 
 import 'Screens/Home/home.dart';
 import 'Screens/onBoarding/onboardingScreen.dart';
 
 int? initScreen;
-void main() async {
+Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
   await Firebase.initializeApp();
   SystemChrome.setSystemUIOverlayStyle(SystemUiOverlayStyle(
@@ -17,12 +17,8 @@ void main() async {
     DeviceOrientation.portraitUp,
     DeviceOrientation.portraitDown,
   ]);
-  SharedPreferences preferences = await SharedPreferences.getInstance();
-  initScreen = preferences.getInt('OnBoardingScreen');
-
   runApp(MyApp());
 }
-
 
 class MyApp extends StatelessWidget {
   static final String title = 'PintSize App';
@@ -33,78 +29,13 @@ class MyApp extends StatelessWidget {
         title: title,
         debugShowCheckedModeBanner: false,
         theme: ThemeData(primaryColor: Colors.white),
-        initialRoute:
-            initScreen == 0 || initScreen == null ? 'onboard' : 'home',
+        initialRoute: "/",
         routes: {
-          'home': (context) => HomeScreen(),
-          'onboard': (context) => OnBoardingScreen()
+          '/': (context) => OnBoardingScreen(),
+          '/home': (context) => HomeScreen(),
+          '/news': (context) => NewsScreen(),
+          // '/donate': (context) => HomeScreen(),
+          // '/profile': (context) => HomeScreen(),
         });
   }
 }
-
-// class MyApp extends StatelessWidget {
-//   static final String title = 'Firebase Setup';
-
-//   @override
-//   Widget build(BuildContext context) => MaterialApp(
-//         debugShowCheckedModeBanner: false,
-//         title: title,
-//         theme: ThemeData(primarySwatch: Colors.blue),
-//         home: MainPage(),
-//       );
-// }
-
-// class MainPage extends StatefulWidget {
-//   @override
-//   _MainPageState createState() => _MainPageState();
-// }
-
-// class _MainPageState extends State<MainPage> {
-//   @override
-//   Widget build(BuildContext context) => Scaffold(
-//         backgroundColor: Colors.indigoAccent.withOpacity(0.2),
-//         appBar: AppBar(
-//           title: Text(MyApp.title),
-//           centerTitle: true,
-//         ),
-//         body: Container(
-//           padding: EdgeInsets.all(16),
-//           child: Center(
-//             child: Column(
-//               mainAxisAlignment: MainAxisAlignment.center,
-//               children: [
-//                 FlutterLogo(size: 100),
-//                 RichText(
-//                   textAlign: TextAlign.center,
-//                   text: TextSpan(
-//                     style: TextStyle(
-//                       fontSize: 36,
-//                       height: 1.3,
-//                       fontWeight: FontWeight.bold,
-//                       color: Colors.black,
-//                     ),
-//                     children: [
-//                       TextSpan(
-//                         text: 'iOS\n',
-//                         style: TextStyle(
-//                           color: Colors.lightGreenAccent.withOpacity(0.8),
-//                           fontWeight: FontWeight.w400,
-//                           height: 3,
-//                         ),
-//                       ),
-//                       TextSpan(
-//                         text: 'Firebase Setup',
-//                         style: TextStyle(
-//                           color: Colors.white,
-//                           fontWeight: FontWeight.bold,
-//                         ),
-//                       ),
-//                     ],
-//                   ),
-//                 ),
-//               ],
-//             ),
-//           ),
-//         ),
-//       );
-// }
