+++ pubspec.yaml
@@ -12,25 +12,28 @@ dependencies:
   flutter:
     sdk: flutter
 
-  cupertino_icons: ^1.0.2
   introduction_screen: ^2.1.0
   shared_preferences: ^2.0.6
   font_awesome_flutter: ^8.8.1
   animations: ^1.1.2
   tuple: ^2.0.0
-  firebase_core: ^1.3.0
   firebase_storage: ^8.1.3
   firebase_auth: ^1.4.1
-  cloud_firestore: ^2.2.2
+  # firebase_analytics: ^8.1.2
 
+
+  cupertino_icons: ^1.0.2
+  firebase_core: ^1.3.0
+  cloud_firestore: ^2.2.2
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
 
+
 flutter:
   uses-material-design: true
 
-  # assets:
-  #   - assets/
+  assets:
+    - assets/
